requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
anthropic==0.40.0
PyGithub==2.3.0
python-dotenv==1.0.1
